Move transaction priority functions into policy
Transaction priority is rules for which low fee transactions
we're willing to include in the blocks we mine. It's policy.
