+++ src/Makefile.am
@@ -139,6 +139,7 @@ BITCOIN_CORE_H = \
   options.h \
   policy/fees.h \
   policy/policy.h \
+  policy/txpriority.h \
   pow.h \
   process_xthinblock.h \
   protocol.h \
@@ -237,6 +238,7 @@ libbitcoin_server_a_SOURCES = \
   options.cpp \
   policy/fees.cpp \
   policy/policy.cpp \
+  policy/txpriority.cpp \
   pow.cpp \
   process_xthinblock.cpp \
   rest.cpp \
