+++ src/coins.h
@@ -290,9 +290,6 @@ public:
     //! Check whether all prevouts of the transaction are present in the UTXO set represented by this view
     bool HaveInputs(const CTransaction& tx) const;
 
-    //! Return priority of tx at height nHeight
-    double GetPriority(const CTransaction &tx, uint32_t nHeight) const;
-
 private:
     CCoinsMap::iterator FetchCoin(const COutPoint &outpoint) const;
