+++ src/qt/coincontroldialog.cpp
@@ -16,6 +16,7 @@
 #include "dstencode.h"
 #include "init.h"
 #include "main.h"
+#include "policy/txpriority.h"
 #include "wallet/wallet.h"
 
 #include <boost/assign/list_of.hpp> // for 'map_list_of()'
